Move manifest API classes
