+++ collect_app/src/main/java/org/odk/collect/android/forms/ServerFormListSynchronizer.java
@@ -1,10 +1,10 @@
 package org.odk.collect.android.forms;
 
-import org.odk.collect.android.logic.ManifestFile;
-import org.odk.collect.android.logic.MediaFile;
 import org.odk.collect.android.openrosa.api.FormAPI;
 import org.odk.collect.android.openrosa.api.FormAPIError;
 import org.odk.collect.android.openrosa.api.FormListItem;
+import org.odk.collect.android.openrosa.api.ManifestFile;
+import org.odk.collect.android.openrosa.api.MediaFile;
 import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.android.utilities.FormDownloader;
