+++ collect_app/src/main/java/org/odk/collect/android/openrosa/api/FormAPI.java
@@ -1,7 +1,5 @@
 package org.odk.collect.android.openrosa.api;
 
-import org.odk.collect.android.logic.ManifestFile;
-
 import java.util.List;
 
 public interface FormAPI {
